Общие изменения по стилистике. Удаление пробелов

+++ src/main/java/ru/yandex/practicum/filmorate/storage/feed/FeedDbStorage.java
@@ -29,21 +29,21 @@ public class FeedDbStorage implements FeedStorage {
     }
 
     @Override
-    public Collection<Event> getFeedById(long userId) {
+    public Collection<Event> getFeedById(int userId) {
         final String sql = "SELECT * FROM feed WHERE user_id = ?";
         log.info("Лента событий пользователя с индентификатором {} отправлена", userId);
         return jdbcTemplate.query(sql, eventMapper, userId);
     }
 
     @Override
-    public Event addEvent(long userId, EventTypes eventType, Operations operation, long entityId) {
+    public Event addEvent(int userId, EventTypes eventType, Operations operation, int entityId) {
         Event event = Event.builder()
                 .timestamp(System.currentTimeMillis())
                 .userId(userId)
                 .eventType(eventType)
                 .operation(operation)
                 .entityId(entityId)
-                .eventId(0L)
+                .eventId(0)
                 .build();
 
         final String sql = "INSERT INTO feed (timestamps, user_id, event_type, operation, entity_id) " +
@@ -53,15 +53,15 @@ public class FeedDbStorage implements FeedStorage {
         jdbcTemplate.update(connection -> {
             PreparedStatement stmt = connection.prepareStatement(sql, new String[]{"event_id"});
             stmt.setLong(1, event.getTimestamp());
-            stmt.setLong(2, event.getUserId());
+            stmt.setInt(2, event.getUserId());
             stmt.setString(3, event.getEventType().toString());
             stmt.setString(4, event.getOperation().toString());
-            stmt.setLong(5, event.getEntityId());
+            stmt.setInt(5, event.getEntityId());
             return stmt;
         }, keyHolder);
 
         if (keyHolder.getKey() != null) {
-            event.setEventId((Long) keyHolder.getKey());
+            event.setEventId((Integer) keyHolder.getKey());
         }
         log.info("Создано событие с индентификатором {} ", event.getEventId());
         return event;
