+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/ReviewMapper.java
@@ -12,7 +12,7 @@ public class ReviewMapper implements RowMapper<Review> {
     @Override
     public Review mapRow(ResultSet rs, int rowNum) throws SQLException {
         return Review.builder()
-                .reviewId(rs.getLong("review_id"))
+                .reviewId(rs.getInt("review_id"))
                 .content(rs.getString("content"))
                 .isPositive(rs.getBoolean("is_positive"))
                 .userId(rs.getInt("user_id"))
